Add files via upload
The objective of the project is to deliver recommendations to the European Hotels Group for expanding their hotel locations in either resort or city settings, in order to best maximize the business's revenue. The workings of the project consist of using various data analysis techniques and models to support the interpreted recommendations, driven by correlational trends between the various parameters present in the data sets.